requirements.txt
pytorch
pytorch-lightning
wandb
omegaconf